Added a library for parsing FBX and used that library to write a method that will take an FBX and strip all physics and collision meshes out of it. This is nesscary because tool.exe has a bug right now that causes it to include physics data in the render jms. Once, or if, that is fixed this can be removed easily as the fbx stripper is it's own static method and is only used in a single place. As for the stripper itself, I tested it against FBX 2014 ASCII and FBX 2014 Binary and it works for both of those. Further testing would be a good idea.

+++ Launcher/Utility/AutoFBX.cs
@@ -5,6 +5,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using ToolkitLauncher.ToolkitInterface;
+using UkooLabs.FbxSharpie;
 
 namespace ToolkitLauncher.Utility
 {
@@ -22,24 +23,40 @@ namespace ToolkitLauncher.Utility
             }
         }
 
-        public static async Task Model(ToolkitBase tk, string path, ModelCompile importType)
+        public static async Task Model(ToolkitBase tk, string path, ModelCompile importType, bool isHalo1)
         {
             /* Check for x.all.fbx in root directory. */
             List<FileInfo> alls = new List<FileInfo>();
             string rootDir = tk.GetDataDirectory() + "\\" + path;
 
-            string[] rootFilePaths = Directory.GetFiles(rootDir, "*.all.fbx");
-            foreach (string f in rootFilePaths)
+            if (!isHalo1) // Don't bother doing this for Halo 1 as the fbx-to-jms tool works differntly for that one
             {
-                FileInfo fi = new FileInfo(f);
-                alls.Add(fi);
-            }
+                string[] rootFilePaths = Directory.GetFiles(rootDir, "*.all.fbx");
+                foreach (string f in rootFilePaths)
+                {
+                    FileInfo fi = new FileInfo(f);
+                    alls.Add(fi);
+                }
 
-            /* Convert x.all.fbx */
-            foreach (FileInfo all in alls)
-            {
-                string outPath = rootDir + "\\" + (all.Name.ToLower().Replace(".all.fbx", ".jms"));
-                await tk.RunTool(ToolType.Tool, new() { "fbx-to-jms", "all", all.FullName, outPath });
+                /* Convert x.all.fbx */
+                foreach (FileInfo all in alls)
+                {
+                    string outPath = rootDir + "\\" + (all.Name.ToLower().Replace(".all.fbx", ".jms"));
+                    await tk.RunTool(ToolType.Tool, new() { "fbx-to-jms", "all", all.FullName, outPath });
+
+                    try
+                    {
+                        string stripInPath = StripFBXCollision(all.FullName);
+                        string stripOutPath = outPath.Replace(".jms", "_render.jms");
+                        await tk.RunTool(ToolType.Tool, new() { "fbx-to-jms", "render", stripInPath, stripOutPath });
+                        if (File.Exists(stripInPath)) { File.Delete(stripInPath); } // Done with this file so delete
+                    }
+                    catch(Exception ex)
+                    {
+                        // Not sure how you want to handle errors of this type.
+                        // If my FBX parse/strip/write/convert fails for whatever reason this catch (should) get it
+                    }
+                }
             }
 
             /* Convert render/collision/physics */
@@ -62,9 +79,9 @@ namespace ToolkitLauncher.Utility
                             string checkPath = typeDir + "\\" + (all.Name.ToLower().Replace(".all.fbx", ".fbx"));
                             if (File.Exists(checkPath)) { continue; } // Skip x.all.fbx if there is an x.fbx with same name in the relevant subfolder
 
-                            if (File.Exists(outPath)) { File.Delete(outPath); }
+                            if (File.Exists(outPath)) { File.Delete(outPath); } // Delete file at outpath because overwrite throws an exception
 
-                            File.Move(inPath, outPath);
+                            if (File.Exists(inPath)) { File.Move(inPath, outPath); } // Move file as long as it exists, a messed up fbx won't output a jms so we check before moving
                         }
                     }
                     /* Convert fbx in relevant sub folders */
@@ -117,6 +134,137 @@ namespace ToolkitLauncher.Utility
                     await tk.RunTool(ToolType.Tool, new() { "fbx-to-jma", fbx.FullName, outPath });
                 }
             }
+        }
+
+        /* Returns path to fbx file with stripped collision and physics data */
+        public static string StripFBXCollision(string path)
+        {
+            FbxDocument fbx = FbxIO.Read(path);
+
+            string FuckYou(UkooLabs.FbxSharpie.Tokens.Token token)
+            {
+                if(token is UkooLabs.FbxSharpie.Tokens.StringToken)
+                {
+                    return ((UkooLabs.FbxSharpie.Tokens.StringToken)token).Value;
+                }
+                if (token is UkooLabs.FbxSharpie.Tokens.Value.IntegerToken)
+                {
+                    return " " + ((UkooLabs.FbxSharpie.Tokens.Value.IntegerToken)token).Value;
+                }
+                if (token is UkooLabs.FbxSharpie.Tokens.Value.LongToken)
+                {
+                    return " " + ((UkooLabs.FbxSharpie.Tokens.Value.LongToken)token).Value;
+                }
+                return "{value}";
+            }
+
+            long FuckYou2(UkooLabs.FbxSharpie.Tokens.Token token)
+            {
+                if (token is UkooLabs.FbxSharpie.Tokens.Value.IntegerToken)
+                {
+                    return ((UkooLabs.FbxSharpie.Tokens.Value.IntegerToken)token).Value;
+                }
+                if (token is UkooLabs.FbxSharpie.Tokens.Value.LongToken)
+                {
+                    return ((UkooLabs.FbxSharpie.Tokens.Value.LongToken)token).Value;
+                }
+                return 0;
+            }
+
+            FbxNode GetFuckYou(FbxDocument f, string identifier)
+            {
+                foreach (FbxNode n in f.Nodes)
+                {
+                    if(n.Identifier.Value.Equals(identifier))
+                    {
+                        return n;
+                    }
+                }
+                return null; // Doooooom
+            }
+
+            /* Step #1 - Search "Objects" for any Models with a name starting with '@' or '$' and then grab the id values for it */
+            List<FbxNode> StNodes5a = new List<FbxNode>();
+            List<long> ModelIDs = new List<long>();
+            foreach(FbxNode node in GetFuckYou(fbx, "Objects").Nodes)
+            {
+                if(node != null && node.Properties != null && node.Properties.Length > 2)
+                {
+                    string name = FuckYou(node.Properties[1]);
+                    long id = FuckYou2(node.Properties[0]);
+                    if (name.StartsWith("Model::@") || name.StartsWith("Model::$"))
+                    {
+                        ModelIDs.Add(id);
+                        continue;
+                    }
+                }
+                StNodes5a.Add(node);
+            }
+
+            /* Step #2 - Search "Connections" for any entries containing the any ModelIDs we collected in property slot #2, store the value from slot #1 as that is the child id for the mesh */
+            List<FbxNode> StNodes6a = new List<FbxNode>();
+            List<long> MeshIDs = new List<long>();
+            foreach (FbxNode node in GetFuckYou(fbx, "Connections").Nodes)
+            {
+                if (node != null && node.Properties != null && node.Properties.Length > 2)
+                {
+                    long id1 = FuckYou2(node.Properties[1]);
+                    long id2 = FuckYou2(node.Properties[2]);
+                    if (ModelIDs.Contains(id2))
+                    {
+                        MeshIDs.Add(id1);
+                        continue;
+                    }
+                }
+                StNodes6a.Add(node);
+            }
+
+            /* Step #3 - Search "Object Definitions" for any entries with the strings 'Geometry' and 'Mesh' in property slots #1 and #2 */
+            List<FbxNode> StNodes5b = new List<FbxNode>();
+            foreach (FbxNode node in StNodes5a)
+            {
+                if (node != null && node.Properties != null && node.Properties.Length > 2)
+                {
+                    long id = FuckYou2(node.Properties[0]);
+                    string slot1 = FuckYou(node.Properties[1]);
+                    string slot2 = FuckYou(node.Properties[2]);
+                    if (slot1.ToLower().Contains("geometry") && slot2.ToLower().Contains("mesh") && MeshIDs.Contains(id))
+                    {
+                        continue;
+                    }
+                }
+                StNodes5b.Add(node);
+            }
+
+            /* Step #4 - Build a new FBX file with the modified nodelist we generated in StNodes */
+            FbxDocument nufbx = new FbxDocument();
+            foreach(FbxNode node in fbx.Nodes)
+            {
+                if (node.Identifier.Value.Equals("Objects"))
+                {
+                    FbxNode nunode = new FbxNode(node.Identifier);
+                    foreach(FbxNode n in StNodes5b) { nunode.AddNode(n); }
+                    foreach (UkooLabs.FbxSharpie.Tokens.Token property in node.Properties) { nunode.AddProperty(property); }
+                    nunode.Value = node.Value;
+                    nufbx.AddNode(nunode);
+                }
+                else if (node.Identifier.Value.Equals("Connections"))
+                {
+                    FbxNode nunode = new FbxNode(node.Identifier);
+                    foreach (FbxNode n in StNodes6a) { nunode.AddNode(n); }
+                    foreach (UkooLabs.FbxSharpie.Tokens.Token property in node.Properties) { nunode.AddProperty(property); }
+                    nunode.Value = node.Value;
+                    nufbx.AddNode(nunode);
+                }
+                else
+                {
+                    nufbx.AddNode(node);
+                }
+            }
+
+            string outPath = path.Replace(".fbx", ".stripped.fbx");
+            FbxIO.WriteAscii(nufbx, outPath);
+            return outPath;
         }
     }
 }
